+++ CHANGELOG.md
@@ -751,3 +751,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Reordered "Delete Data on Uninstall" to be placed just above the "Clear App Data" setting and just below the "Open Kivixa Folder" setting.
   
 ---
+## [0.8.34] - 2026-05-16
+
+### Changed
+- Version bump to 0.8.34
+
+---